code.
In this code, the EBike class represents an e-bike. It has private member variables for battery level and speed. The class provides member functions such as accelerate() to increase the speed, brake() to decrease the speed, chargeBattery() to charge the battery level to 100%, and displayStatus() to show the current battery level and speed.

The main() function demonstrates the usage of the e-bike by creating an instance of EBike, calling various member functions, and displaying the status of the e-bike at different points.

Feel free to modify the code and add additional functionalities as per your requirements for the e-bike simulation.